+++ test/lang_extract_test.exs
@@ -1,0 +1,5 @@
+defmodule LangExtractTest do
+  use ExUnit.Case, async: true
+
+  doctest LangExtract
+end